requirements.txt
# EconSearch requirements
fastapi
uvicorn[standard]
pydantic
pypdf
openai
scikit-learn
numpy
python-dotenv
